requirements.txt
asgiref==3.6.0
black==23.1.0
boto3==1.26.72
botocore==1.29.72
click==8.1.3
Django==4.1.6
django-cors-headers==3.13.0
django-environ==0.9.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
jmespath==1.0.1
Markdown==3.4.1
mypy-extensions==0.4.3
packaging==23.0
pathspec==0.11.0
Pillow==9.4.0
platformdirs==2.6.2
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.7.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.14
whitenoise==6.3.0
